modularize records code
